@@ -26,7 +26,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  intl: ^0.17.0
+  intl: ^0.18.0
   flutter_image_compress: ^1.0.0
   percent_indicator: ^3.4.0
   font_awesome_flutter: ^9.0.0
@@ -46,42 +46,38 @@ dependencies:
   hive: ^2.0.4
   hive_flutter: ^1.1.0
   maps_launcher: ^2.0.0
-  path_provider: ^2.0.1
+  path_provider: ^2.1.2
   share_plus: ^3.0.4
   timeago: '3.1.0'
   shared_preferences: ^2.0.6
   location: ^4.3.0
-  # flutter_facebook_login: ^3.0.0
   flutter_facebook_auth: ^3.5.3
-  google_sign_in: ^5.2.1
+  google_sign_in: ^6.1.6
   image_cropper: ^1.3.0
-  flutter_paystack: ^1.0.4+1
+  flutter_paystack: ^1.0.5
   url_launcher: ^6.0.6
   random_string: ^2.1.0
   crypt: ^4.0.1
-  sentry_flutter: ^6.0.1
-  video_trimmer: ^0.6.0
-  file_picker: ^4.2.0
+  sentry_flutter: ^7.18.0
+  video_trimmer: ^3.0.1
+  file_picker: ^5.3.1
   video_player: ^2.1.6
-  video_thumbnail: ^0.4.3
+  video_thumbnail: ^0.5.3
   visibility_detector: ^0.2.0
-  # cached_video_player: ^2.0.0-dev.1
   socket_io_client: ^1.0.1
-  firebase_core: "^1.4.0"
-  firebase_messaging: "^10.0.4"
+  firebase_core: "^2.24.2"
+  firebase_messaging: "^14.7.1"
   flutter_uploader: ^3.0.0-beta.1
-  awesome_notifications: ^0.6.13
+  awesome_notifications: ^0.9.3+1
   need_resume: ^1.0.6
   video_compress: ^3.1.0
   photofilters: ^3.0.1
   collection: ^1.15.0-nullsafety.4
-  cloud_firestore: ^3.1.4
-  firebase_analytics: ^9.0.0-dev.3
-  google_mobile_ads: ^0.13.6
+  cloud_firestore: ^4.14.0
+  firebase_analytics: ^10.8.0
   huawei_push: ^5.3.0+304
   google_api_availability: ^3.0.1
   provider: ^6.0.1
-  # light_compressor: ^1.2.1
 
 dev_dependencies:
   flutter_test:
@@ -90,6 +86,9 @@ dev_dependencies:
   hive_generator: ^1.1.0
   build_runner: ^2.0.6
   flutter_native_splash: ^1.3.1
+
+dependency_overrides:
+  intl: ^0.17.0-nullsafety.2
   
 
 # For information on the generic Dart part of this file, see the
