@@ -189,8 +189,10 @@ class _ClubEventsPageState extends State<ClubEventsPage> {
               decoration: BoxDecoration(
                   color: Colors.deepPurple,
                   borderRadius: BorderRadius.circular(900)),
-              child: FlatButton(
+              child: TextButton(
+                style: TextButton.styleFrom(
                 padding: EdgeInsets.only(left: 8),
+              ),
                 onPressed: () {
                   Navigator.pop(context);
                 },
@@ -272,8 +274,7 @@ class _ClubEventsPageState extends State<ClubEventsPage> {
               decoration: BoxDecoration(
                   color: Colors.deepPurple,
                   borderRadius: BorderRadius.circular(9)),
-              child: FlatButton(
-                padding: EdgeInsets.zero,
+              child: TextButton(
                 onPressed: () {
                   _scrollController.animateTo(
                     0.0,
