@@ -44,7 +44,6 @@ class ClubCollapsingAppBar extends SliverPersistentHeaderDelegate {
       height: (expandedHeight - 36),
       child: Stack(
         fit: StackFit.expand,
-        overflow: Overflow.visible,
         children: [
           Positioned.fill(
             child: CachedNetworkImage(
@@ -116,8 +115,7 @@ class ClubCollapsingAppBar extends SliverPersistentHeaderDelegate {
                   color: Colors.deepPurple,
                   borderRadius: BorderRadius.circular(9),
                 ),
-                child: FlatButton(
-                  padding: EdgeInsets.zero,
+                child: TextButton(
                   onPressed: () => onFavButtonClick(),
                   child: Icon(
                     favouritesClubsCubit.checkClubExists(clubID)
@@ -149,8 +147,10 @@ class ClubCollapsingAppBar extends SliverPersistentHeaderDelegate {
               color: Colors.deepPurple,
               borderRadius: BorderRadius.circular(900),
             ),
-            child: FlatButton(
-              padding: EdgeInsets.only(left: 8),
+            child: TextButton(
+              style: TextButton.styleFrom(
+                padding: EdgeInsets.only(left: 8),
+              ),
               onPressed: () {
                 Navigator.pop(context);
               },
@@ -195,8 +195,7 @@ class ClubCollapsingAppBar extends SliverPersistentHeaderDelegate {
           shape: CircleBorder(),
           color: Colors.deepPurple,
           child: Container(
-            child: FlatButton(
-              padding: EdgeInsets.zero,
+            child: TextButton(
               onPressed: () {
                 MapsLauncher.launchCoordinates(latitude!, longitude!);
               },
