@@ -274,8 +274,10 @@ class _FollowingPageState extends State<FollowingPage> {
             decoration: BoxDecoration(
                 color: Colors.deepPurple,
                 borderRadius: BorderRadius.circular(900)),
-            child: FlatButton(
-              padding: EdgeInsets.only(left: 8),
+            child: TextButton(
+              style: TextButton.styleFrom(
+                padding: EdgeInsets.only(left: 8),
+              ),
               onPressed: () {
                 Navigator.pop(context);
               },
@@ -318,8 +320,7 @@ class _FollowingPageState extends State<FollowingPage> {
               decoration: BoxDecoration(
                   color: Colors.deepPurple,
                   borderRadius: BorderRadius.circular(9)),
-              child: FlatButton(
-                padding: EdgeInsets.zero,
+              child: TextButton(
                 onPressed: () {
                   _scrollController.animateTo(
                     0.0,
