@@ -215,7 +215,7 @@ class ActionButton extends StatelessWidget {
           color: theme.colorScheme.secondary,
           elevation: 4.0,
           child: IconTheme.merge(
-            data: theme.accentIconTheme,
+            data: theme.iconTheme,
             child: IconButton(
               onPressed: onPressed,
               icon: icon!,
